Rename "Version" to "Full spec" at the top
This confused me in the past; it's not super-clear that "Version" goes
to the full spec.

Also make the dropdown a bit wider, so all the text fits without
breaking on my system (would be even better to automatically size it,
but I didn't feel like fighting this Tailwind thing).