Add date to spec, translate "Text version"
This adds the date to the top of every spec. I also noticed the "Text
version" was always in English, so translate that too.

(I machine translated the texts, but DeepL and Google Translate seem to
agree, and a quick search seems to verify they're not horrible wrong, so
should be good).

Just keep a manual list of the dates. Otherwise it's spending half an
hour or more to automate about 1 minute of work every few years.

Fixes #12